feat: 입력/삭제 예외 및 확인 메세지

+++ static/js/components/TodoDelete.js
@@ -1,13 +1,14 @@
-import {EVENT_TYPE, STATE} from "../utils/constants.js";
+import {EVENT_TYPE, STATE, MESSAGE} from "../utils/constants.js";
 
 export function TodoDelete({onDelete}) {
     const $todo_list = document.querySelector("#todo-list");
 
     const onDeleteHandler = event => {
         if (event.target.className === "destroy") {
-            event.target.parentElement.parentElement.className = STATE.COMPLETED;
-            const index = event.target.parentElement.parentElement.getAttribute("data-index");
-            onDelete(index);
+            if (confirm(MESSAGE.CONFIRM_DELETE)) {
+                const index = event.target.parentElement.parentElement.getAttribute("data-index");
+                onDelete(index);
+            }
         }
     };
